requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mysql-connector-python==8.0.23
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.17.4
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.15.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
Send2Trash==1.5.0
six==1.13.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
